+++ BaseVisitor.hpp
@@ -5,6 +5,6 @@ class BaseVisitor
 {
   public:
     void makeCurrent();
-    virtual void visitNumber(int number) {}
+    virtual void visitNumber(float number) {}
     virtual void visitOperator(char op) {}
 };
